pyproject.toml
[tool.poetry]
name = "aio-openapi"
version = "2.8.0"
description = "Minimal OpenAPI asynchronous server application"
documentation = "https://aio-openapi.readthedocs.io"
repository = "https://github.com/quantmind/aio-openapi"
authors = ["Luca <luca@quantmind.com>"]
license = "BSD-3-Clause"
readme = "readme.md"
packages = [
    { include = "openapi" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: JavaScript",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Framework :: AsyncIO",
    "Environment :: Web Environment",
]

[tool.poetry.urls]
repository = "https://github.com/quantmind/aio-openapi"
issues = "https://github.com/quantmind/aio-openapi/issues"


[tool.poetry.dependencies]
python = ">=3.8,<4"
aiohttp = "^3.8.0"
httptools = "^0.3.0"
simplejson = "^3.17.5"
asyncpg = "^0.25.0"
SQLAlchemy = "^1.4.27"
SQLAlchemy-Utils = "^0.38.2"
alembic = "psycopg2-binary"
psycopg2-binary = "^2.9.2"
click = "^8.0.3"
python-dateutil = "^2.8.2"
email-validator = "^1.1.3"
pytz = "^2021.3"
PyYAML = "^6.0"
uvloop = {version = "^0.16.0", optional = true}
aiodns = {version = "^3.0.0", optional = true}
PyJWT = {version = "^2.3.0", optional = true}
colorlog = {version = "^6.6.0", optional = true}
phonenumbers = {version = "^8.12.37", optional = true}
cchardet = {version = "^2.1.7", optional = true}
Sphinx = {version = "^4.3.0", optional = true}
sphinx-copybutton = {version = "^0.4.0", optional = true}
sphinx-autodoc-typehints = {version = "^1.12.0", optional = true}
aiohttp-theme = {version = "^0.1.6", optional = true}
recommonmark = {version = "^0.7.1", optional = true}

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
pytest = "^6.2.5"
flake8-blind-except = "^0.2.0"
flake8-builtins = "^1.5.3"
flake8-commas = "^2.1.0"
codecov = "^2.1.12"
coverage = "^6.1.2"
coveralls = "^3.3.1"
mypy = "^0.930"
sqlalchemy-stubs = "^0.4"
sentry-sdk = "^1.4.3"
python-dotenv = "^0.19.2"
openapi-spec-validator = "^0.3.1"
pytest-cov = "^3.0.0"
pytest-aiohttp = "^0.3.0"
pytest-mock = "^3.6.1"
isort = "^5.10.1"
types-pytz = "^2021.3.6"
types-simplejson = "^3.17.5"
types-python-dateutil = "^2.8.11"
factory-boy = "^3.2.1"

[tool.poetry.extras]
dev = ["aiodns", "PyJWT", "colorlog", "phonenumbers", "cchardet"]
docs = [
    "Sphinx",
    "recommonmark",
    "aiohttp-theme",
    "sphinx-copybutton",
    "sphinx-autodoc-typehints",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
